+++ src/components/burger-ingredients/burger-ingredients.js
@@ -7,7 +7,7 @@ import burgerIngridients from './burger-ingredients.module.css';
 import {
   Tab
 } from '@ya.praktikum/react-developer-burger-ui-components';
-import IngredientCard from '../ingredient-card/ingredient-card';
+import IngredientCard from './ingredient-card/ingredient-card';
 import Modal from '../modal/modal';
 import IngredientDetails from '../ingredient-details/ingredient-details';
